+++ azure-storage-blob/ChangeLog.md
@@ -2,6 +2,12 @@
 
 > See [BreakingChanges](BreakingChanges.md) for a detailed list of API breaks.
 
+## Version 1.2.0rc1:
+
+- Support for 2017-11-09 REST version. Please see our REST API documentation and blog for information about the related added features.
+- Support for write-once read-many containers.
+- Added support for OAuth authentication for HTTPS requests(Please note that this feature is available in preview).
+
 ## Version 1.1.0:
 
 - Support for 2017-07-29 REST version. Please see our REST API documentation and blogs for information about the related added features.