+++ azure-storage-common/ChangeLog.md
@@ -2,7 +2,7 @@
 
 > See [BreakingChanges](BreakingChanges.md) for a detailed list of API breaks.
 
-## Version XX.XX.XX:
+## Version 1.2.0rc1:
 
 - Increased default socket timeout to a more reasonable number for Python 3.5+.
 - Fixed bug where seekable streams (request body) were not being reset for retries.